Move cty-repl-2 to `cty repl`.
